+++ R/emm.R
@@ -47,7 +47,7 @@ emm <- function(model, ci.lvl = .95, type = c("fe", "re", "fe.zi", "re.zi", "sim
   # check class of fitted model
   fun <- get_predict_function(model)
   # check model family, do we have count model?
-  faminfo <- get_model_info(model)
+  faminfo <- .get_model_info(model)
 
   # compute predictions here
   preds <-