+++ R/get_predictions_zeroinfl.R
@@ -43,9 +43,9 @@ get_predictions_zeroinfl <- function(model, fitfram, ci.lvl, linv, type, fun, ty
   if (type == "fe.zi") {
 
     mf <- insight::get_data(model)
-    clean_terms <- get_clear_vars(terms)
+    clean_terms <- .get_cleaned_terms(terms)
 
-    newdata <- get_expanded_data(
+    newdata <- .get_data_grid(
       model,
       mf,
       terms,